rename inference to likelihood
