+++ flamedisx/__init__.py
@@ -3,7 +3,7 @@ __version__ = '0.2.2'
 from .utils import *
 from .xenon_corrections import *
 from .source import *
-from .inference import *
+from .likelihood import *
 
 from .er_nr_base import *
 from .x1t_sr0 import *
